added feature file egress unhappy path scenarios
